Reorganizing some tests, and adding a base test class
